Refactor Crews into separate app & include PK
