+++ dcs/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = (
 
 
     # Internal Application
+    'crew',
     'multimedia',
 )
